Implement PINFakeBasicAuth
This commit implements an extra option to increase compatability for backend authentication on a MFA-enabled proxy server. By using this option, the PIN portion of a password will be added to a Basic Authentication header in the HTTP request.

This is intended to be used where PIN authentication is shared between the proxy and the back-end application (e.g. SQL, LDAP, Kerberos).